requirements.txt
requests==2.20.0
PyExecJS==1.5.1